requirements.txt
diffusers
accelerate
transformers
sentencepiece
protobuf
einops
torchvision
matplotlib
safetensors
peft
controlnet_aux